factor out CSS hook
