+++ packages/application/src/hooks/useComponents.ts
@@ -1,14 +1,17 @@
 import type { ComponentCompilerResponse } from '@bos-web-engine/compiler';
-import { useCallback, useEffect, useRef, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 
 import { UseComponentsParams } from '../types';
 
 /**
  * Provides an interface for managing containers
+ * @param appendStylesheet callback to add container stylesheets to the Component tree CSS
+ * @param compiler Web Engine compiler instance
  * @param config parameters to be applied to the entire Component tree
  * @param rootComponentPath Component path for the root Component
  */
 export function useComponents({
+  appendStylesheet,
   compiler,
   config,
   rootComponentPath,
@@ -20,37 +23,6 @@ export function useComponents({
   const [rootComponentSource, setRootComponentSource] = useState<string | null>(
     null
   );
-
-  const containerStylesheet = useRef<CSSStyleSheet | null>(null);
-
-  useEffect(() => {
-    const style = document.createElement('style');
-    style.id = `bwe-styles-${Date.now()}`;
-    style.appendChild(document.createTextNode(''));
-    document.head.appendChild(style);
-
-    // @ts-expect-error StyleSheetList can be inlined despite TS complaints about [Symbol.iterator]()
-    containerStylesheet.current = [...document.styleSheets].find(
-      ({ ownerNode }) => ownerNode === style
-    );
-  }, []);
-
-  const appendStylesheet = useCallback((containerStyles: string) => {
-    const css = new CSSStyleSheet();
-    css.replaceSync(containerStyles);
-
-    // @ts-expect-error StyleSheetList can be inlined despite TS complaints about [Symbol.iterator]()
-    for (let { cssText } of css.cssRules) {
-      containerStylesheet.current!.insertRule(cssText);
-    }
-  }, []);
-
-  const resetContainerStylesheet = useCallback(() => {
-    const rulesCount = containerStylesheet.current!.cssRules.length;
-    for (let i = rulesCount - 1; i >= 0; i--) {
-      containerStylesheet.current!.deleteRule(i);
-    }
-  }, [containerStylesheet]);
 
   const addComponent = useCallback((componentId: string, component: any) => {
     setComponents((currentComponents) => ({
@@ -169,7 +141,6 @@ export function useComponents({
     error,
     hooks,
     getComponentRenderCount,
-    resetContainerStylesheet,
     setComponents,
   };
 }
